Renamed inference classes. Added doc comments. Optimised aspects of bot inference classes
